Move test model to folder tests
